requirements.txt
kafka-python
pyspark
SQLAlchemy
psycopg2
confluent-kafka
py-ccloud
fastavro
avro-python3